fix(board): make the home lens the baseline, dedupe lens copy
Addresses code-review findings on the boardDefaultLens PR:

- Filtered-state chrome now keys off homeLens, not the literal All: the
  lens-menu trigger variant and both `isFiltered` predicates (filter bar +
  saved views) compare against homeLens, so a non-All home no longer shows a
  permanent "Clear filters" link / filled trigger on its own untouched
  landing. "Clear filters" returns to home.
- handlePosted no longer bounces every post to /board/all: it stays put when
  the current view already contains a fresh self-post (unfiltered all/mine —
  a self-authored conversation is isMine), and returns to the All surfacing
  baseline only from a status/memo lens or a scoped view, where the post
  might not match. Honors "your own action always surfaces" without throwing
  a mine-home author off their home.
- INV-33: extract BOARD_LENS_DEFS to lib/board/lens-defs.ts as the single
  source of lens copy; the settings "Board home" control reads label/
  description from it instead of duplicating the strings.
